generate genesis block on start

+++ ubuntu-setup/cleanup.sh
@@ -2,9 +2,12 @@
 
 function Main() {
     rm -rf /var/lib/sawtooth/* &&
-    rm -rf /var/log/sawtooth/* &&
-    rm -rf "${HOME}"/.sawtooth/keys/*{priv,pub} &&
-    rm -rf /etc/sawtooth/keys/*{priv,pub}
+        rm -rf /var/log/sawtooth/*
+
+    if [[ "${1}" == "keys" ]]; then
+        rm -rf "${HOME}"/.sawtooth/keys/*{priv,pub} &&
+            rm -rf /etc/sawtooth/keys/*{priv,pub}
+    fi
 
     return $?
 }
